#1 refactor: 옵션 선택 컴포넌트 파일 병합
